+++ nym-vpn-app/src/screens/node/index.ts
@@ -1,2 +1,3 @@
 export { default as NodeEntry } from './NodeEntry';
+export { default as QuicTag } from './QuicTag';
 export * from './details';